pyproject.toml
[project]
name = "boreal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.23.0",
    "azure-search-documents>=11.5.2",
    "fastapi>=0.115.12",
    "ipykernel>=6.29.5",
    "langchain-community>=0.3.24",
    "langchain-core>=0.3.60",
    "langchain-openai>=0.3.17",
    "langchain-text-splitters>=0.3.8",
    "langgraph>=0.4.5",
    "langsmith[otel]>=0.3.42",
    "openai>=1.79.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "crawl4ai>=0.6.3",
    "fastapi-cli>=0.0.7",
]
